refactor createTake to audioVM
