+++ common/src/main/kotlin/org/wycliffeassociates/otter/common/domain/content/TakeCreator.kt
@@ -1,0 +1,52 @@
+/**
+ * Copyright (C) 2020-2024 Wycliffe Associates
+ *
+ * This file is part of Orature.
+ *
+ * Orature is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * Orature is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Orature.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package org.wycliffeassociates.otter.common.domain.content
+
+import org.wycliffeassociates.otter.common.audio.wav.IWaveFileCreator
+import org.wycliffeassociates.otter.common.data.primitives.MimeType
+import org.wycliffeassociates.otter.common.data.workbook.Take
+import java.io.File
+import java.time.LocalDate
+import javax.inject.Inject
+
+class TakeCreator @Inject constructor(
+    private val waveFileCreator: IWaveFileCreator
+)  {
+
+    fun createNewTake(
+        newTakeNumber: Int,
+        filename: String,
+        audioDir: File,
+        createEmpty: Boolean
+    ): Take {
+        val takeFile = audioDir.resolve(File(filename))
+        val newTake = Take(
+            name = takeFile.name,
+            file = takeFile,
+            number = newTakeNumber,
+            format = MimeType.WAV,
+            createdTimestamp = LocalDate.now()
+        )
+        if (createEmpty) {
+            newTake.file.createNewFile()
+            waveFileCreator.createEmpty(newTake.file)
+        }
+        return newTake
+    }
+}
